fix: Correct layout when user logs out
