+++ src/app/login/page.jsx
@@ -1,18 +1,14 @@
-import { EmailField, PasswordField } from '../../components/Shared/form-fields.jsx';
 import Link from 'next/link';
+import Login from '../../components/login';
 
 export const metadata = {
     title: 'Login',
 }
 
-export default function Login() {
+export default function LoginPage() {
     return (
         <div className="auth">
-            <form action="/auth/login" method="POST">
-                <EmailField />
-                <PasswordField />
-                <button type="submit">Log In</button>
-            </form>
+            <Login />
             <Link href="/forgot-password">Forgot password?</Link>
             <Link href="/register">Sign up</Link>
         </div>
