+++ src/components/login.jsx
@@ -1,0 +1,13 @@
+import { EmailField, PasswordField } from './Shared/form-fields';
+
+export default function Login() {
+    return (
+        <>
+            <form action="/auth/login" method="POST">
+                <EmailField />
+                <PasswordField />
+                <button type="submit">Log In</button>
+            </form>
+        </>
+    );
+}
