+++ src/user-provider.jsx
@@ -12,6 +12,8 @@ export default function UserProvider({ children }) {
     const [user, setUser] = useState(null);
     const value = { user, setUser };
 
+    console.log("context rendering...")
+
     useEffect(() => {
         const fetchUser = async () => {
             const supabase = createClientComponentClient();
